Backend/requirements.txt
fastapi[standard]>=0.116.1,<0.117.0
pydantic>=2.11.9,<3.0.0
passlib[argon2]>=1.7.4,<2.0.0
pymongo>=4.15.1,<5.0.0
bcrypt>=4.3.0,<5.0.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6

python-dotenv>=1.0.1
argon2_cffi>=23.1.0,<24.0.0
bcrypt>=4.3.0,<5.0.0
python-multipart>=0.0.6
pydantic[email]>=2.11.9,<3.0.0